gauge fixes

+++ relayer/src/metrics/para.rs
@@ -81,8 +81,8 @@ impl MetricsCollector {
 
             let address = parachain_runtime::storage().beefy_mmr().beefy_authorities();
             self.update_metric(&address, block_hash, |vset| {
-                metrics::gauge!(PARA_BEEFY_CURRENT_ID, vset.id);
-                metrics::gauge!(PARA_BEEFY_CURRENT_LEN, vset.len as u64);
+                metrics::gauge!(PARA_BEEFY_CURRENT_ID, vset.id as f64);
+                metrics::gauge!(PARA_BEEFY_CURRENT_LEN, vset.len as f64);
             })
             .await;
 
@@ -91,8 +91,8 @@ impl MetricsCollector {
                 .beefy_next_authorities();
 
             self.update_metric(&address, block_hash, |vset| {
-                metrics::gauge!(PARA_BEEFY_NEXT_ID, vset.id);
-                metrics::gauge!(PARA_BEEFY_NEXT_LEN, vset.len as u64);
+                metrics::gauge!(PARA_BEEFY_NEXT_ID, vset.id as f64);
+                metrics::gauge!(PARA_BEEFY_NEXT_LEN, vset.len as f64);
             })
             .await;
 
@@ -107,10 +107,10 @@ impl MetricsCollector {
                         .beefy_light_client()
                         .current_validator_set(network_id);
                     self.update_metric(&address, block_hash, |vset| {
-                        metrics::gauge!(PARA_BEEFY_LIGHT_CLIENT_CURRENT_ID, vset.id, labels);
+                        metrics::gauge!(PARA_BEEFY_LIGHT_CLIENT_CURRENT_ID, vset.id as f64, labels);
                         metrics::gauge!(
                             PARA_BEEFY_LIGHT_CLIENT_CURRENT_LEN,
-                            vset.len as u64,
+                            vset.len as f64,
                             labels
                         );
                     })
@@ -120,8 +120,8 @@ impl MetricsCollector {
                         .beefy_light_client()
                         .next_validator_set(network_id);
                     self.update_metric(&address, block_hash, |vset| {
-                        metrics::gauge!(PARA_BEEFY_LIGHT_CLIENT_NEXT_ID, vset.id, labels);
-                        metrics::gauge!(PARA_BEEFY_LIGHT_CLIENT_NEXT_LEN, vset.len as u64, labels);
+                        metrics::gauge!(PARA_BEEFY_LIGHT_CLIENT_NEXT_ID, vset.id as f64, labels);
+                        metrics::gauge!(PARA_BEEFY_LIGHT_CLIENT_NEXT_LEN, vset.len as f64, labels);
                     })
                     .await;
